refactor: Re-structured library to make it instalable
The scripts were moved to a sub-folder. A setup
script was created to make the library instalable.
import statements in some scripts were updated to
match the new structure of the library. The mat file
is not referenced with absolute path to avoid
errors when using the library in other projects.