refactor: Rolled-back imports in modules
In commit 255101e, importing modules was done considering
that the library was going to be installed to be used in
other projects. However, this posses a problem when using
conda environments. When exporting the dependencies of the
environment, the library will be part of the process.
However, when trying to create a new environemnt with the
library using the resulting yml file. The process will fail
because the library is not part of any repository.

Therefore, using the library in other projects requires a
hack to make the code available in the system. This is not
difficult to do.